requirements.txt
Flask==0.9
Flask-Login==0.1.3
Jinja2==2.6
PIL==1.1.7
Werkzeug==0.8.3
distribute==0.6.24
httplib2==0.8
mock==1.0.1
mongokit==0.8.3
mongomock==0.0.2
nose==1.2.1
py-bcrypt==0.3
pymongo==2.5
pyotp==1.3.1
sentinels==0.0.6
six==1.3.0
twilio==3.4.5
unittest2==0.5.1
wsgiref==0.1.2
